refactor(ErrorResponse): ProblemDetail 상속을 없애고 ErrorCode에 대응되는 응답으로 구현
- `ProblemDetail`을 상속하면서 발생하는 `type`, `title`, `detail`, `instance` 필드를 반환하지 않도록 상속 제거
- 변경된 `ErrorCode`에 맞추어 `code`, `status`, `message`만 반환하도록 변경
